RST firewalled tcp connections
A closer look reveals that core.tcpConn.Close isn't doing the
desired clean-up; and probably a core.tcpConn.Abort is a better
match to firewall a connection cleanly.

Calling core.tcpConn.Abort is better left to the downstream
caller. Returning a error from registered-tcp-handler does
in fact abort the connection as required.

Similar change for udp connections too, only that instead of
both closing the connection and returning an error, we err
on the side of just returning an error.

Ocassional SIGSEGVs originating deep in LwIP when firewalling
and de-firewalling connections hopefully now is put to bed.