SOCKS5 over TCP, DNS Proxy, and DNSCrypt over TCP
DNSCrypt v1 and v2 with support for anonymous relays. Currently, only
DNSCrypt over TCP has been implemented, that is it forwards both UDP and TCP
DNS requests over TCP. No DNSCrypt support for Tor and DoH. DoH is anyway
supported outside of DNSCrypt with the only difference that the reference
implementation (upstream) also works with user-supplied TLS certs.

Support for a tonne of dnscrypt-proxy (upstream) configuration is also left
unimplemented in favour of simplicity in terms of code and opinionated
defaults which aren't necessarily any better: for example, even when
anonymous relays are set, the client always fetches certificates
from the resolver and not via the relays which means client identity leaks
to the resolver. Another feature left out is selecting resolvers based on
round-trip time (latency) because it involved adding a write-lock in the
DNS query path and that's unfortunate and so the entire thing was removed.
The resolvers are instead selected by power-of-2 algorithm instead.

Forwarding requests to a SOCK5 endpoint is supported only for TCP and
that's due to a limitation in the underlying x/net/proxy package.

DNS Proxy forwards plain-old unencrypted DNS over TCP and DNS over UDP payload
to any reachable TCP and UDP endpoint respectively. It, hopefully, correctly
maps UDP connections to UDP endpoint and TCP connections to TCP endpoint
whilst preserving source and destination IP addresses (NAT'ing). This also
sets up the code nicely for a future split-horizon DNS implementation.