On device blocking and other API changes
On device blocking using the code ported over to golang from bravedns resolver
which is in javascript. CNAME cloaking and wildcard prefix matching works
as well as it does on the resolver. This is supported for both dns over https
and dnscrypt, but not for dns forwarding proxy.

The compressed radix-trie and associated rank-directory, trie-configuration and
blocklists-metadata are needed for the blocking to happen on-device. These are
downloaded and supplied by the kotlin layer. The radix-trie is fronted by a LFU
cache to soften the blow of repeated checks for a single query sometimes that
take more than 10ms per check. LFU pretty much reduces cached response times
to nearly 1ms.

If the remote resolver is bravedns, the blocklist stamp sent in x-nile-flags
http-header is decoded using blocklist-metadata file (again, to be supplied
by the kotlin layer) to retrieve the actual blocklist names and sent over
to the kotlin layer through the listener.

Piggy backing on this is a refactor in dnscrypt code-base of moving out some
common code to the xdns package. Confusingly, dnsx package implements on-device
blocking and facilitates parsing x-nile-flags header for blocks by the
bravedns resolver.

APIs changes in tunnel/intra.go to make it saner for the kotlin layer to
set various proxies (dns, dnscrypt, and socks5) and reason about them, though
one inconsistency remains: socks5 requires tun-mode to be set to socks before
starting the socks5 proxy whereas for every other proxy, it is the other way
around: that is, proxies need to be started before changing the tun-mode. The
latter behaviour prevents connection leaks and so, socks5 would need to be
eventually moved to being consistent with the rest of the proxy APIs.