Too many unknown conns on Android 9 and below
On Android 9 and below, the app makes use of /proc/net to figure out owner
uid given a 5-tuple. It turns out that often times v6 procfs ends up with
entries for v4 conns (happy-eyeballs) but the app never accounted for those.

Now, even when connections are v4, make an effort to check the v6 procfs
entries to see if a v6-only network is tunneling v4 packets.

Additionally, keep a cache of procfs entires for 30s or so to reduce the
very costly scans on procfs.