procfs cache backed by sync.Map
A couple of time it was observed that the cache was over-run with race
conditions which results in a runtime panic in latest go versions.

Even though a switch to sync.Map is not the most optimal, it at least
guarantees no race related failures in the wild. Uptime triumphs all and
this was a quick win versus sanitizing the cache for races, even if sync.Map
API isn't type-safe as such. Well, good news is, generics in go are imminent.