fix blocklists reports; jit name-resolution for doh
NewDohTransport bubbles up exceptions when the hostname couldn't be resolved.
The current contract means the error is essentially treated as unrecoverable by
client code, when in fact, it isn't. Instead, the code now retries lazily /
just-in-time to resolve IPs for the doh hostname.

Blocklists (shadowed variable) was reported as "empty" to the listener in all
cases; whilst query-error was set to BadQuery whenever a DNS request was
blocked locally. Pesky lil bugs with straight-forward fixes.