dnsx/alg: support a second underlying dns transport
Results from the default (primary) transport is used for alg, whilst the
response from secondary transport (if set) is used for making "block" / "allow"
decisions by the client code. That is, ips in response to a domain-name from
the secondary is merged with ips from primary. If the secondary responds with
unspecified ips (0.0.0.0 / ::), then the client code may use that information
to make a decision on whether to allow or block the resulting outgoing internet
connection for that ALG'd domain name.