dnsx/cacher: del corrupt cache entries when detected
At times, answer for another query, y, is cached against unrelated query, x.
This is undeseriable since an invalid DNS answer induces a loss of internet
connectivity like behaviour in client apps.

At the time of processing cached answers, check if the requested question
matches up to the answer. If not, delete the entry, and proceed to fetch
a new answer from upstream.

This patch is merely fixing the symptom and not the root-cause (which has
been hard to track down).