dnsx/wall: if blockstamp is invalid, let local blocker take over
sometimes, the base64 blockstamp actually represents blocklists themselves
(when summaries are populated from caches or the alg transport), and in those
cases, fallback onto local blocklists to see if the answer must be blocked,
that is, IPs set to 0.0.0.0 or :: as the incoming  answer might be a resolved
answer (that is, not 0.0.0.0 or ::). This answer when sent to the client,
results in a TCP/UDP connection to IPs in the resolved answer. The firewall
must then decide (depending on presence of blocklists against that domain)
whether to allow / block the request.

In cases where TCP/UDP firewall must decide whether to block a domain or not,
for example when a domain is trusted by one or two apps but blocked globally,
this "may be blocked [by the firewall]" behaviour is fine. In other cases,
it is more efficient to simply block at the DNS layer with 0.0.0.0 or ::

Falling back onto local blocklists helps mitigate cases where the base64
blockstamp is set but is invalid AND the resolved answer is also left
unmodified by whoever sent the blockstamp (usually, caches), resulting in an
undesirable and avoidable "may be blocked [by the firewall]" behaviour.