rnet: servers
servers listen on all underlying interfaces and accept incoming connections
or packets and forward them to the gateway. If a L4 hop is set, then conns
are forwarded to that L4 hop. ipn.Proxy is an example of a hop.

Anything that can connect to these rnet.servers over LAN (for example,
Browsers on the same network as the Android running Rethink) can use
any of the L4 ipn.Proxy transports (like WireGuard, Http CONNECT, SOCKS5,
and the RPN over WebSockets) for egress (just like EveryProxy does).

The initial implementation (untested) contains support for SOCKS5 servers.