gomobile: use protect.Conn as protect.PacketConn cannot be bound
gomobile can only bind fns that return 2 values (and the 2nd value must
be an error). protect.PacketConn.ReadFrom returns 3 values and so gomobile
refuses to generate code. use protect.Conn as a stand-in but assert the
returned value back to *net.UDPConn (which confirms to protect.PacketConn)
before use.