dnsx: Preset transport synthesizes IP answers
A new preset transport takes into account NsOpt.IPCSV to synthesize answers
for A, AAAA queries, and substitutes ip hints in HTTPS, SVCB records.

As a special case, if IPCSV has unspecified IP, it synthesizes a blocked
response regardless of query type.

If Preset transport fails to make a case for synthesis (A query but none of the
preset IPs are ip4, or AAAA query but none are ip6 etc), it forwards the
query to a real DNS transport.