dnsx/resolver: unset errs if answer non-empty
In some cases, the transports both error out (EOF) and send a response; and
in such cases (EOF denotes a force close), errors should not override the
actual answer, that is, it being overlooked by caching and alg layers.

This results in rest of the engine behaving as if these queries were never
resolved (neither cached nor mapped) when in fact they were.

A wrapper func, dnsx.Req, unset err if returned answer is not empty.