dnsx: do not remove when adding new transports
dnscrypt servers are already added to underlying DcProxy multi-transport
by the Resolver.Add() is called by intra.AddDNSCryptTransport. And so,
removing transports before adding one with the same ID results in the latest
DcProxy transport being removed. That is, AddDNSCryptTransport adds a new
DNSCrypt transport to DcProxy which is followed by Resolver.Add()... which
incorrectly removes the DNSCrypt transport that was just added to DcProxy.