dnsx,ipmapper: empty string ids inside transport-id lists
If no DNS transport-id is sent, prefer dnsx.Default to avoid being caught in a
loop where a user-seet DNS, say dnsx.Preferred is chosen to resolve a domain,
but then, to resolve dnsx.Preferred, we need to choose a DNS transport-id,
which of course shouldn't be dnsx.Preferred again; or we end up in a catch-22.

That is, in case no DNS transport-id is specified, the only safe choice is
dnsx.Default.