dnsx/alg: add secondary answer ips in all cases
tcp,udp,icmp flow events rely on presence of unspecified ipv4/ipv6 addrs
to know if a connection must be blocked (ie, applying a domain-name rule
at the time of the flow event). So, in all cases, and especially in cases
where the domain is blocked (either by primary or secondary or on-device),
the ips returned by querying secondary must always contain ips, if any.