dnsx: on Android R+, socket ownership on DNS queries is retained
And so, the UID as got from SocketListener.Flow (from Kotlin/Java)
is likely to be of the owner UID that actually sent the DNS request.

On Android Q and below, a combination of ALG+dnsx.Fixed is needed
to unmask the true owner of a DNS request by mapping made up unique IPs
to domain names at resolution time, but then actually resolving the
domain names at TCP/UDP connection time (and thereby, unmasking the owner UID
as on Android M+, ownership information of TCP/UDP sockets is not "hidden".