dnsx/transport: reimpl refresh
Re-adding the same transport results in the older transport (which is the
same as the current one we wish to re-add) being "stopped". This results
in transport being unusable ever after a single Refresh call.

Since dialers.Clear already clears up cached IPs of DNS Transport hostnames,
use Cacher.Clear to surgically clear just the caches and not rely on "re-adds"
to nuke the the cachers up.