dnsx: always call into choose healthy proxy
ProxyProvider.ProxyTo fn also ensures any idling / sleeping proxy is
woken up before client code uses it for dialing. As such, when the DNS
transport has a fixed "relay" (which is a proxy), it also needs to go through
ProxyProvider.ProxyTo fn so that it may be woken up if it was asleep / idling.