dnsx: do not hold a strong ref to relay proxy
relay proxy must always be got from the supplied ProxyProvider, like with all
other proxies that the DNS transport may use; and must not be stored a strong
reference to as the lifecycle of proxies is not under the control of
these DNS Transports. For instance, any proxy (all of which are keyed against
a string ID), could be changed, updated, or stopped on-the-fly by ipn.Proxies.