netstack: prevent stop() on invalid fd
If stop() is called twice (which it shouldn't be but it somehow is), the
2nd stop() call would invariably act on an "invalidfd" (currently set to -1).
The code today is setup to panic when that happens. Instead change it to
an error so that the runtime does not crash.

Calling stop() on an invalidfd shouldn't happen unless firestack is in Loopback
mode, which isn't enabled by default anyway. That is, in Loopback, the client
code is required to call tunnel.Dispose before tunnel.SetLink. tunnel.Dispose
stop()s the current fd and sets the new fd to "invalidfd" until the next
tunnel.SetLink call sets a proper (tunnel) fd.