intra/tunnel: hide actual tunnel impl from client
With the Restart() impl, the underlying tunnel can be swapped out for a new one
in which case the client would have to access it in a thread-safe way over
a mutex, say, but it can't since the underlying tunnel's exposed as a top-level
construct by intra/tunnel.

Also, move to core.Volatile over rw mutexes for the underlying tun and remove
a bunch of obsolete funcs that are not longer in use.