tcp,udp,icmp: guard verbose/debug logs
some logs are in the critical path, that is, exec on every packet flow and so
they cause connection slow downs. The Golang logger is not lightweight and
acquires mutexes on every write. Firestack's logger also does a bunch of
bookkeeping to prevent spammy logs and calls up the runtime to print
stacktraces.